mcp/requirements.txt
fastapi==0.115.7
uvicorn==0.35.0
python-multipart==0.0.20
mcp-python==0.1.4
unstructured==0.18.13
sentence-transformers==5.1.0
psycopg2-binary==2.9.10
minio==7.2.16
python-dotenv==1.1.1
numpy==2.3.2